Corrected the video Height & Width
The original `height` and `width` values were set to a very small *horizontal* (landscape) frame size. By setting the `height` to `auto` and the `width` to `75%` (instead of fixed values), the video will play in the correct *vertical* (portrait) fame size, and it'll be larger and much easier to see in the Acode App.